+++ tuxedomask/repositories/__init__.py
@@ -1,9 +1,9 @@
 # -*- coding: utf-8 -*-
 
 from .base_repository import EntityConflict, EntityNotFound, BaseRepository
-from .tuxedo_mask_base_repository import TuxedoMaskBaseRepository
-from .tuxedo_mask_applications_repository import TuxedoMaskApplicationsRepository
-from .tuxedo_mask_users_repository import TuxedoMaskUsersRepository
+from .tuxedomask_base_repository import TuxedoMaskBaseRepository
+from .tuxedomask_applications_repository import TuxedoMaskApplicationsRepository
+from .tuxedomask_users_repository import TuxedoMaskUsersRepository
 
 __all__ = ['BaseRepository',
            'EntityConflict',