+++ tuxedomask/repositories/tests/test_tuxedo_mask_base_repository.py
@@ -6,7 +6,7 @@ from nose.tools import (assert_is_instance,
                         assert_true,
                         raises)
 
-from tuxedo_mask import repositories
+from tuxedomask import repositories
 
 
 class MockEntity: