+++ tuxedomask/resources/login_attempts_resource.py
@@ -4,7 +4,7 @@ import flask
 import flask_restful
 
 from . import authentication
-from tuxedo_mask import views
+from tuxedomask import views
 
 
 class LoginAttemptsResource(flask_restful.Resource):