+++ tuxedomask/services/__init__.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 
 from .base_service import BaseService
-from .tuxedo_mask_service import TuxedoMaskService
+from .tuxedomask_service import TuxedoMaskService
 
 __all__ = ['BaseService', 'TuxedoMaskService']