+++ tuxedomask/views/users_view.py
@@ -7,7 +7,7 @@ import marshmallow
 from marshmallow import fields, validate
 
 from . import BaseView
-from tuxedo_mask import models
+from tuxedomask import models
 
 
 class ContainsAtLeast(validate.Validator):